Temporary fixing

+++ src/main/java/com/group3/ezquiz/repository/OptionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.group3.ezquiz.model.Option;
 
-public interface OptionRepository extends JpaRepository<Option, Integer> {
+public interface OptionRepository extends JpaRepository<Option, Long> {
 
 }
